### 2016-07-21 #43 @whedon
Submission: PyGBe: Python, GPUs and Boundary elements for biomolecular electrostatics
**Submitting author:** @labarba (<a href="http://orcid.org/0000-0001-5812-2711">Lorena A. Barba</a>)
**Repository:** <a href="https://github.com/barbagroup/pygbe" target ="_blank">https://github.com/barbagroup/pygbe</a>
**Version:** 2.1
