### 2016-07-22 869a676
Fixing bibtex entry.
