### 2016-07-25 62fec1f
Added citation to PyCUDA
