+++ paper.bib
@@ -77,3 +77,19 @@
     Volume = {247},
     Year = {2013}}
 
+@article{kloeckner_pycuda_2012,
+   author = {{Kl{\"o}ckner}, Andreas
+        and {Pinto}, Nicolas
+        and {Lee}, Yunsup
+        and {Catanzaro}, B.
+        and {Ivanov}, Paul
+        and {Fasih}, Ahmed },
+   title = "{PyCUDA and PyOpenCL: A Scripting-Based Approach to GPU Run-Time Code Generation}",
+   journal = "Parallel Computing",
+   volume = "38",
+   number = "3",
+   pages = "157--174",
+   year = "2012",
+   issn = "0167-8191",
+   doi = "10.1016/j.parco.2011.09.001",
+}
