### 2016-07-25 cdea5da
Added reference to PyCUDA and Barnes & Hut
