+++ paper.md
@@ -29,8 +29,8 @@ It computes solvation energies for proteins modeled with any number of dielectri
 The mathematical formulation follows @YoonLenhoff1990 for solving the Poisson-Boltzmann equation of the [implicit-solvent](https://en.wikipedia.org/wiki/Implicit_solvation) model in integral form.
 
 PyGBe achieves both algorithmic and hardware acceleration.
-The solution algorithm uses a [Barnes-Hut](https://en.wikipedia.org/wiki/Barnes–Hut_simulation) treecode to accelerate each iteration of a GMRES solver to O(N logN), for N unknowns. 
-It exploits NVIDIA GPU hardware on the most computationally intensive parts of the code using CUDA kernels in the treecode, interfacing with PyCUDA. 
+The solution algorithm uses a [Barnes-Hut](https://en.wikipedia.org/wiki/Barnes–Hut_simulation) treecode from @BarnesHut1986 to accelerate each iteration of a GMRES solver to O(N logN), for N unknowns. 
+It exploits NVIDIA GPU hardware on the most computationally intensive parts of the code using CUDA kernels in the treecode, interfacing with PyCUDA (see @kloeckner_pycuda_2012). 
 Some parts of the code are written in C++, wrapped using SWIG. 
 
 # References
