### 2016-07-25 80d8647
update all print statements
