+++ pygbe/main.py
@@ -246,11 +246,11 @@ def main(argv=sys.argv, log_output=True, return_output_fname=False):
     if log_output:
         sys.stdout = Logger(os.path.join(output_dir, outputfname))
     ### Time stamp
-    print 'Run started on:'
-    print '\tDate: %i/%i/%i' % (timestamp.tm_year, timestamp.tm_mon,
-                                timestamp.tm_mday)
-    print '\tTime: %i:%i:%i' % (timestamp.tm_hour, timestamp.tm_min,
-                                timestamp.tm_sec)
+    print('Run started on:')
+    print('\tDate: {}/{}/{}'.format(timestamp.tm_year, timestamp.tm_mon,
+                                timestamp.tm_mday))
+    print('\tTime: {}:{}:{}'.format(timestamp.tm_hour, timestamp.tm_min,
+                                timestamp.tm_sec))
     TIC = time.time()
 
     print('Config file: {}'.format(configFile))
@@ -292,8 +292,8 @@ def main(argv=sys.argv, log_output=True, return_output_fname=False):
             param.Neq += N_aux
         else:
             param.Neq += 2 * N_aux
-    print '\nTotal elements : %i' % param.N
-    print 'Total equations: %i' % param.Neq
+    print('\nTotal elements : {}'.format(param.N))
+    print('Total equations: {}'.format(param.Neq))
 
     results_dict['total_elements'] = param.N
     results_dict['N_equation'] = param.Neq
@@ -312,14 +312,14 @@ def main(argv=sys.argv, log_output=True, return_output_fname=False):
         kernel = 1
 
     ### Generate interaction list
-    print 'Generate interaction list'
+    print('Generate interaction list')
     tic = time.time()
     generateList(surf_array, field_array, param)
     toc = time.time()
     list_time = toc - tic
 
     ### Transfer data to GPU
-    print 'Transfer data to GPU'
+    print('Transfer data to GPU')
     tic = time.time()
     if param.GPU == 1:
         dataTransfer(surf_array, field_array, ind0, param, kernel)
@@ -329,7 +329,7 @@ def main(argv=sys.argv, log_output=True, return_output_fname=False):
     timing = Timing()
 
     ### Generate RHS
-    print 'Generate RHS'
+    print('Generate RHS')
     tic = time.time()
     if param.GPU == 0:
         F = generateRHS(field_array, surf_array, param, kernel, timing, ind0)
@@ -340,22 +340,22 @@ def main(argv=sys.argv, log_output=True, return_output_fname=False):
     rhs_time = toc - tic
 
     setup_time = toc - TIC
-    print 'List time          : %fs' % list_time
-    print 'Data transfer time : %fs' % transfer_time
-    print 'RHS generation time: %fs' % rhs_time
-    print '------------------------------'
-    print 'Total setup time   : %fs\n' % setup_time
+    print('List time          : {}s'.format(list_time)
+    print('Data transfer time : {}s'.format(transfer_time)
+    print('RHS generation time: {}s'.format(rhs_time)
+    print('-'*30)
+    print('Total setup time   : {}s'.format(setup_time)})
 
     tic = time.time()
 
     ### Solve
-    print 'Solve'
+    print('Solve')
     phi = numpy.zeros(param.Neq)
     phi = gmres_mgs(surf_array, field_array, phi, F, param, ind0, timing,
                        kernel)
     toc = time.time()
     solve_time = toc - tic
-    print 'Solve time        : %fs' % solve_time
+    print('Solve time        : {}s'.format(solve_time))
[…]
