### 2016-07-25 e029627
update print statements and simplify conditional
Lines 111-118 appear to be a direct copy of the previous conditional
block with the only difference being whether something is a dirichlet or
neumann surface.  I collapsed those two into a single conditional to
simplify the code.  I also found a few spots that might have trouble
with integer division?  Probably not, but good to be safe.
