### 2016-07-25 f38284c
update print statements to py3 syntax using format str
