+++ pygbe/output.py
@@ -17,8 +17,8 @@ def print_summary(surf_array, field_array, param, results_dict):
     """
 
     Nsurf = len(surf_array)
-    print 28 * '-' + '\n'
-    print '%i surfaces:\n' % Nsurf
+    print(28 * '-' + '\n')
+    print('{} surfaces:\n'.format(Nsurf))
     for i in range(len(surf_array)):
         N_aux = len(surf_array[i].triangle)
         rr = numpy.zeros(len(surf_array[i].tree))
@@ -26,52 +26,56 @@ def print_summary(surf_array, field_array, param, results_dict):
             rr[ii] = surf_array[i].tree[ii].r
         Levels = numpy.log(surf_array[i].tree[0].r/numpy.min(rr))/numpy.log(2) + 1
 
-        print 'Surface %i:'%i
-        print '\t%i elements'%(N_aux)
-        print '\tSurface type       : '+surf_array[i].surf_type
-        print '\tCells              : %i'%len(surf_array[i].tree)
-        print '\tTwigs              : %i'%len(surf_array[i].twig)
-        print '\tLevels             : %i'%Levels
-        print '\tC0 size            : %f'%surf_array[i].tree[0].r
-        print '\tC0 box center      : %f, %f, %f'%(surf_array[i].tree[0].xc, surf_array[i].tree[0].yc, surf_array[i].tree[0].zc)
-        print '\tTwig cell size     : %f'%(numpy.min(rr))
-        print '\tRbox/theta         : %f'%(numpy.min(rr)/param.theta)
-        print '\tAnalytic distance  : %f'%(numpy.average(numpy.sqrt(2*surf_array[i].Area))/param.threshold)
-        print '\tElem. per sq Ang   : %f'%(1/numpy.average(surf_array[i].Area))
-        print '\tMax, min, avg elem.: %s, %s, %s'%(numpy.max(surf_array[i].Area),numpy.min(surf_array[i].Area),numpy.average(surf_array[i].Area))
-        print '\tTotal area         : %f'%(numpy.sum(surf_array[i].Area))
+        print('Surface {}:'.format(i))
+        print('\t{} elements'.format(N_aux))
+        print('\tSurface type       : {}'.format(surf_array[i].surf_type))
+        print('\tCells              : {}'.format(len(surf_array[i].tree)))
+        print('\tTwigs              : {}'.format(len(surf_array[i].twig)))
+        print('\tLevels             : {}'.format(Levels))
+        print('\tC0 size            : {}'.format(surf_array[i].tree[0].r))
+        print('\tC0 box center      : {}, {}, {}'.format(surf_array[i].tree[0].xc,
+                                                         surf_array[i].tree[0].yc,
+                                                         surf_array[i].tree[0].zc))
+        print('\tTwig cell size     : {}'.format(numpy.min(rr)))
+        print('\tRbox/theta         : {}'.format(numpy.min(rr)/param.theta))
+        print('\tAnalytic distance  : {}'.format(numpy.average(numpy.sqrt(2*surf_array[i].Area))/param.threshold))
+        print('\tElem. per sq Ang   : {}'.format(1/numpy.average(surf_array[i].Area)))
+        print('\tMax, min, avg elem.: {}, {}, {}'.format(numpy.max(surf_array[i].Area),
+                                                         numpy.min(surf_array[i].Area),
+                                                         numpy.average(surf_array[i].Area)))
+        print('\tTotal area         : {}'.format(numpy.sum(surf_array[i].Area)))
 
-    print 28*'-'+'\n'
+    print(30*'-'+'\n')
 
     Nfield = len(field_array)
-    print '%i regions:\n'%Nfield
+    print('{} regions:\n'.format(Nfield))
     for i in range(len(field_array)):
[…]
