### 2016-07-25 05eb222
delete commented code and update print statements
