### 2016-07-25 698cdb7
remove unused commented code
