### 2016-07-25 514af70
fix a few missing parens in new print statements
