### 2016-07-25 53165df
update subprocess returns for python 3
in py2, things are returned as strings, in py3, they're returned as
bytes, so the `decode()` method needs to be invoked
