### 2016-07-25 59aa18b
add explicit import to semianalytical.py
I have no idea why this ever worked on Python2, but it definitely
doesn't on Python3 without a proper import
