### 2016-07-25 a5e4ec8
change use of `file` as object opener
no longer valid in Python 3, all instances have been changed to use a
context manaager so we don't have multiple open file objects hanging
around
