### 2016-07-25 aa7b2c7
force integer division for calculating indices
Py3 doesn't like floats as indices and since floating point division is
the default, we need to force integer division here.
