+++ pygbe/tree/FMMutils.py
@@ -383,10 +383,10 @@ def sortPoints(surface, Cells, twig, param):
     surface.xjSort = surface.xj[surface.sortSource]
     surface.yjSort = surface.yj[surface.sortSource]
     surface.zjSort = surface.zj[surface.sortSource]
-    surface.AreaSort = surface.Area[surface.sortSource / param.K]
-    surface.sglInt_intSort = surface.sglInt_int[surface.sortSource / param.K]
-    surface.sglInt_extSort = surface.sglInt_ext[surface.sortSource / param.K]
-    surface.triangleSort = surface.triangle[surface.sortSource / param.K]
+    surface.AreaSort = surface.Area[surface.sortSource // param.K]
+    surface.sglInt_intSort = surface.sglInt_int[surface.sortSource // param.K]
+    surface.sglInt_extSort = surface.sglInt_ext[surface.sortSource // param.K]
+    surface.triangleSort = surface.triangle[surface.sortSource // param.K]
 
 
 def computeIndices(P, ind0):
