### 2016-07-25 d9be93b
remove bare exception, those are bad
if you have an exception that you catch, make sure it is specifically
the exception you are trying to catch, otherwise other things will gum
up the works
