### 2016-07-25 d273e18
fix a few more floating point indices
