### 2016-07-26 a0f2532
make 'Levels' an explicit int
