### 2016-07-26 e6c3ba9
fix readData for loop
I under-indented this bit so the charges weren't being properly read
from the pqr files.  I think the logic in this section can be improved
and simplified, but for now it appears to be working.
