+++ pygbe/util/readData.py
@@ -202,33 +202,24 @@ def readpqr(filename, REAL):
         for line in lines:
             line = numpy.array(line.split())
 
-        line_aux = []
+            line_aux = []
 
-        if line[0] == 'ATOM':
-            for l in range(len(line) - 6):
-                aux = line[5 + len(line_aux)]
-                if len(aux) > 14:
-                    X = readCheck(aux, REAL)
-                    for i in range(len(X)):
-                        line_aux.append(X[i])
-#                        line_test.append(str(X[i]))
-                else:
-                    #                    line_test.append(line[5+len(line_aux)])
-                    line_aux.append(REAL(line[5 + len(line_aux)]))
+            if line[0] == 'ATOM':
+                for l in range(len(line) - 6):
+                    aux = line[5 + len(line_aux)]
+                    if len(aux) > 14:
+                        X = readCheck(aux, REAL)
+                        for i in range(len(X)):
+                            line_aux.append(X[i])
+                    else:
+                        line_aux.append(REAL(line[5 + len(line_aux)]))
 
-#            line_test.append(line[len(line)-1])
-            x = line_aux[0]
-            y = line_aux[1]
-            z = line_aux[2]
-            q.append(line_aux[3])
-            pos.append([x, y, z])
+                x = line_aux[0]
+                y = line_aux[1]
+                z = line_aux[2]
+                q.append(line_aux[3])
+                pos.append([x, y, z])
 
-#           for i in range(10):
-#                f.write("%s\t"%line_test[i])
-#            f.write("\n")
-
-#    f.close()
-#    quit()
     pos = numpy.array(pos)
     q = numpy.array(q)
     Nq = len(q)
