### 2016-07-26 95c3c8c
force integer division when calculating parameter 'Nm'
