### 2016-07-26 #43 @arfon
@kyleniemeyer - the paper has just been updated for this submission.
