### 2016-07-26 7530d8c
update readme and remove trailing whitespace
added notes on the flags for config and param files.

also removed trailing whitespace because trailing whitespace is horrible.
