+++ README.md
@@ -1,35 +1,35 @@
 # PyGBe: Python, GPUs and Boundary elements for biomolecular electrostatics
 
-PyGBe—pronounced _pigbē_—is a Python code to apply the boundary element method for molecular-electrostatics 
+PyGBe—pronounced _pigbē_—is a Python code to apply the boundary element method for molecular-electrostatics
 calculations in a continuum model.
-It computes solvation energies for proteins modeled with any number of dielectric regions. 
+It computes solvation energies for proteins modeled with any number of dielectric regions.
 The mathematical formulation follows Yoon and Lenhoff (1990) for solving the Poisson-Boltzmann equation of the [implicit-solvent](https://en.wikipedia.org/wiki/Implicit_solvation) model in integral form.
 
 PyGBe achieves both algorithmic and hardware acceleration.
-The solution algorithm uses a [Barnes-Hut](https://en.wikipedia.org/wiki/Barnes–Hut_simulation) treecode to accelerate each iteration of a GMRES solver to O(N logN), for N unknowns. 
-It exploits NVIDIA GPU hardware on the most computationally intensive parts of the code using CUDA kernels in the treecode, interfacing with PyCUDA. 
-Some parts of the code are written in C++, wrapped using SWIG. 
+The solution algorithm uses a [Barnes-Hut](https://en.wikipedia.org/wiki/Barnes–Hut_simulation) treecode to accelerate each iteration of a GMRES solver to O(N logN), for N unknowns.
+It exploits NVIDIA GPU hardware on the most computationally intensive parts of the code using CUDA kernels in the treecode, interfacing with PyCUDA.
+Some parts of the code are written in C++, wrapped using SWIG.
 
 ## Installation
 
-The following instructions assume that the operating system is Ubuntu. Run the 
+The following instructions assume that the operating system is Ubuntu. Run the
 corresponding commands in your flavor of Linux to install.
 
 ### Dependencies (last tested)
 * Python 2.7.11
 * Numpy 1.11.0
 * SWIG 3.0.8
-* NVCC 7.0 
+* NVCC 7.0
     * gcc < 4.10
 * PyCUDA 2016.1.1
 
-#### Python and Numpy 
+#### Python and Numpy
 
 To install the specific version of these packages we recommend using either [conda](http://conda.pydata.org/docs/get-started.html) or [pip](http://python-packaging-user-guide.readthedocs.org/en/latest/installing/).
 
 #### SWIG
 
-To install SWIG we recommend using either `conda`, your distribution package manager or [SWIG's website](http://www.swig.org/download.html).  
+To install SWIG we recommend using either `conda`, your distribution package manager or [SWIG's website](http://www.swig.org/download.html).
 
 #### NVCC
 
@@ -48,12 +48,12 @@ We summarize the commands to install PyCUDA on Ubuntu here:
     > cd pycuda-2015.1.3
     > python configure.py --cuda-root=/usr/local/cuda
     > make
-    > sudo make install 
-    
+    > sudo make install
+
 If you are not installing PyCUDA systemwide, do not use `sudo` to install and simply run
 
     > make install
-    
+
 as the final command.
 
 Test the installation by running the following:
@@ -61,7 +61,7 @@ Test the installation by running the following:
     > cd test
     > python test_driver.py
 
-PyGBe has been run and tested on Ubuntu 12.04, 13.10, 15.04 and 16.04 (using gcc < 4.10). 
+PyGBe has been run and tested on Ubuntu 12.04, 13.10, 15.04 and 16.04 (using gcc < 4.10).
 
 ### Installing PyGBe
 
@@ -71,20 +71,20 @@ Create a clone of the repository on your machine:
[…]
