### 2016-08-01 eb3906b
begin moving surface functions into surface classes
If a function that currently resides in surface.py applies
exclusively (or near exclusively) to a surface, then it should be a
method within the surface class.  This makes for much better control
flow and also eliminates the need for explicitly typing out a bunch of
arguments.
