+++ pygbe/classes.py
@@ -1,6 +1,13 @@
 import numpy
 import time
 
+from pygbe.tree.FMMutils import addSources, sortPoints, generateTree, findTwigs
+from pygbe.tree.direct import computeDiagonal
+from pygbe.util.semi_analytical import GQ_1D
+from pygbe.util.readData import (readVertex, readTriangle, readpqr, readcrd,
+                           readFields, read_surface)
+from pygbe.quadrature import quadratureRule_fine, getGaussPoints
+
 class Event():
     """
     Class for logging like in pycuda's cuda.Event()
@@ -127,90 +134,230 @@ class Surface():
     """
 
     def __init__(self):
-        self.triangle = []  # indices to triangle vertices
-        self.vertex   = []  # position of vertices
-        self.XinV     = []  # weights input for single layer potential
-        self.XinK     = []  # weights input for double layer potential
-        self.Xout_int = []  # output vector of interior operators
-        self.Xout_ext = []  # output vector of exterior operators
-        self.xi       = []  # x component of center
-        self.yi       = []  # y component of center
-        self.zi       = []  # z component of center
-        self.xj       = []  # x component of gauss nodes
-        self.yj       = []  # y component of gauss nodes
-        self.zj       = []  # z component of gauss nodes
-        self.Area     = []  # Area of triangles
-        self.normal   = []  # normal of triangles
-        self.sglInt_int = []  # singular integrals for V for internal equation
-        self.sglInt_ext = []  # singular integrals for V for external equation
-        self.xk       = []  # position of gauss points on edges
-        self.wk       = []  # weight of gauss points on edges
-        self.Xsk      = []  # position of gauss points for near singular integrals
-        self.Wsk      = []  # weight of gauss points for near singular integrals
-        self.tree     = []  # tree structure
+#        self.triangle = []  # indices to triangle vertices
+#        self.vertex   = []  # position of vertices
+#        self.XinV     = []  # weights input for single layer potential
+#        self.XinK     = []  # weights input for double layer potential
+#        self.Xout_int = []  # output vector of interior operators
+#        self.Xout_ext = []  # output vector of exterior operators
+#        self.xi       = []  # x component of center
+#        self.yi       = []  # y component of center
+#        self.zi       = []  # z component of center
+#        self.xj       = []  # x component of gauss nodes
+#        self.yj       = []  # y component of gauss nodes
+#        self.zj       = []  # z component of gauss nodes
+#        self.Area     = []  # Area of triangles
+#        self.normal   = []  # normal of triangles
+#        self.sglInt_int = []  # singular integrals for V for internal equation
+#        self.sglInt_ext = []  # singular integrals for V for external equation
+#        self.xk       = []  # position of gauss points on edges
+#        self.wk       = []  # weight of gauss points on edges
+#        self.Xsk      = []  # position of gauss points for near singular integrals
+#        self.Wsk      = []  # weight of gauss points for near singular integrals
+#        self.tree     = []  # tree structure
         self.twig     = []  # tree twigs
-        self.xiSort   = []  # sorted x component of center
-        self.yiSort   = []  # sorted y component of center
-        self.ziSort   = []  # sorted z component of center
-        self.xjSort   = []  # sorted x component of gauss nodes
-        self.yjSort   = []  # sorted y component of gauss nodes
[…]
