+++ pygbe/surface.py
@@ -107,141 +107,6 @@ def zeroAreas(s, triangle_raw, Area_null):
     return Area_null
 
 
-def fill_surface(surf, param):
-    """
-    It fills the surface with all the necessary information to solve it.
-
-    -It sets the Gauss points.
-    -It generates tree, computes the indices and precompute terms for M2M.
-    -It generates preconditioner.
-    -It computes the diagonal integral for internal and external equations.
-
-    Arguments
-    ----------
-    surf     : class, surface that we are studying.
-    param    : class, parameters related to the surface we are studying.
-
-    Returns
-    --------
-    time_sort: float, time spent in sorting the data needed for the treecode.
-    """
-
-    N = len(surf.triangle)
-    Nj = N * param.K
-    # Calculate centers
-    surf.xi = numpy.average(surf.vertex[surf.triangle[:], 0], axis=1)
-    surf.yi = numpy.average(surf.vertex[surf.triangle[:], 1], axis=1)
-    surf.zi = numpy.average(surf.vertex[surf.triangle[:], 2], axis=1)
-
-    surf.normal = numpy.zeros((N, 3))
-    surf.Area = numpy.zeros(N)
-
-    L0 = surf.vertex[surf.triangle[:, 1]] - surf.vertex[surf.triangle[:, 0]]
-    L2 = surf.vertex[surf.triangle[:, 0]] - surf.vertex[surf.triangle[:, 2]]
-    surf.normal = numpy.cross(L0, L2)
-    surf.Area = numpy.sqrt(surf.normal[:, 0]**2 + surf.normal[:, 1]**2 +
-                           surf.normal[:, 2]**2) / 2
-    surf.normal[:, 0] = surf.normal[:, 0] / (2 * surf.Area)
-    surf.normal[:, 1] = surf.normal[:, 1] / (2 * surf.Area)
-    surf.normal[:, 2] = surf.normal[:, 2] / (2 * surf.Area)
-
-    # Set Gauss points (sources)
-    surf.xj, surf.yj, surf.zj = getGaussPoints(surf.vertex, surf.triangle,
-                                               param.K)
-
-    x_center = numpy.zeros(3)
-    x_center[0] = numpy.average(surf.xi).astype(param.REAL)
-    x_center[1] = numpy.average(surf.yi).astype(param.REAL)
-    x_center[2] = numpy.average(surf.zi).astype(param.REAL)
-    dist = numpy.sqrt((surf.xi - x_center[0])**2 + (surf.yi - x_center[1])**2 +
-                      (surf.zi - x_center[2])**2)
-    R_C0 = max(dist)
-
-    # Generate tree, compute indices and precompute terms for M2M
-    surf.tree = generateTree(surf.xi, surf.yi, surf.zi, param.NCRIT, param.Nm,
-                             N, R_C0, x_center)
-    C = 0
-    surf.twig = findTwigs(surf.tree, C, surf.twig, param.NCRIT)
-
-    addSources(surf.tree, surf.twig, param.K)
-    #    addSources3(surf.xj,surf.yj,surf.zj,surf.tree,surf.twig)
-    #    for j in range(Nj):
-    #        C = 0
-    #        addSources2(surf.xj,surf.yj,surf.zj,j,surf.tree,C,param.NCRIT)
-
-    surf.xk, surf.wk = GQ_1D(param.Nk)
-    surf.Xsk, surf.Wsk = quadratureRule_fine(param.K_fine)
-
-    # Generate preconditioner
-    # Will use block-diagonal preconditioner (AltmanBardhanWhiteTidor2008)
-    #If we have complex dielectric constants we need to initialize Precon with
[…]
