### 2016-08-01 1742325
break out more class methods in Surface
