+++ pygbe/classes.py
@@ -155,7 +155,7 @@ class Surface():
 #        self.Xsk      = []  # position of gauss points for near singular integrals
 #        self.Wsk      = []  # weight of gauss points for near singular integrals
 #        self.tree     = []  # tree structure
-        self.twig     = []  # tree twigs
+#        self.twig     = []  # tree twigs
 #        self.xiSort   = []  # sorted x component of center
 #        self.yiSort   = []  # sorted y component of center
 #        self.ziSort   = []  # sorted z component of center
@@ -217,6 +217,7 @@ class Surface():
 #        self.XskDev     = []
 #        self.WskDev     = []
 #        self.kDev       = []
+        self.twig = []
 
     def fill_surface(self, param):
         """
@@ -244,32 +245,120 @@ class Surface():
 
         self.calc_norms()
         # Set Gauss points (sources)
-        self.xj, self.yj, self.zj = getGaussPoints(self.vertex, self.triangle,
-                                                param.K)
+        self.getGaussPoints(param.K)
 
-        x_center = numpy.zeros(3)
-        x_center[0] = numpy.average(self.xi).astype(param.REAL)
-        x_center[1] = numpy.average(self.yi).astype(param.REAL)
-        x_center[2] = numpy.average(self.zi).astype(param.REAL)
-        dist = numpy.sqrt((self.xi - x_center[0])**2 + (self.yi - x_center[1])**2 +
-                        (self.zi - x_center[2])**2)
-        R_C0 = max(dist)
+        #x_center = numpy.zeros(3)
+        #x_center[0] = numpy.average(self.xi).astype(param.REAL)
+        #x_center[1] = numpy.average(self.yi).astype(param.REAL)
+        #x_center[2] = numpy.average(self.zi).astype(param.REAL)
+        #dist = numpy.sqrt((self.xi - x_center[0])**2 + (self.yi - x_center[1])**2 +
+        #                (self.zi - x_center[2])**2)
+        #R_C0 = max(dist)
+        self.calc_distance(param)
 
         # Generate tree, compute indices and precompute terms for M2M
         self.tree = generateTree(self.xi, self.yi, self.zi, param.NCRIT, param.Nm,
-                                self.N, R_C0, x_center)
+                                self.N, self.R_C0, self.x_center)
         C = 0
         self.twig = findTwigs(self.tree, C, self.twig, param.NCRIT)
 
         addSources(self.tree, self.twig, param.K)
-        #    addSources3(self.xj,self.yj,self.zj,self.tree,self.twig)
-        #    for j in range(Nj):
-        #        C = 0
-        #        addSources2(self.xj,self.yj,self.zj,j,self.tree,C,param.NCRIT)
 
         self.xk, self.wk = GQ_1D(param.Nk)
         self.Xsk, self.Wsk = quadratureRule_fine(param.K_fine)
 
+        self.generate_preconditioner()
+
+        tic = time.time()
+        sortPoints(self, self.tree, self.twig, param)
+        toc = time.time()
+        time_sort = toc - tic
+
+        return time_sort
+
+    def calc_centers(self):
+        self.xi = numpy.average(self.vertex[self.triangle[:], 0], axis=1)
+        self.yi = numpy.average(self.vertex[self.triangle[:], 1], axis=1)
+        self.zi = numpy.average(self.vertex[self.triangle[:], 2], axis=1)
+
+    def calc_norms(self):
+        self.normal = numpy.zeros((self.N, 3))
+        self.Area = numpy.zeros(self.N)
+
[…]
