### 2016-08-01 6ee2c55
prefer array operations on a per-array basis
