### 2016-08-01 94a242d
rename gauss function and simplify
