+++ pygbe/quadrature.py
@@ -479,55 +479,3 @@ def getWeights(K):
 
     return w
     # yapf: enable
-
-def getGaussPoints(y, triangle, n):
-    """
-    It gets the Gauss points for far away integrals.
-
-    Arguments
-    ----------
-    y       : list, vertices of the triangles.
-    triangle: list, indices for the corresponding triangles.
-    n       : int (1,3,4,7), desired Gauss points per element.
-
-    Returns
-    --------
-    xi[:,0] : position of the gauss point in the x axis.
-    xi[:,1] : position of the gauss point in the y axis.
-    xi[:,2] : position of the gauss point in the z axis.
-    """
-
-    N  = len(triangle) # Number of triangles
-    xi = numpy.zeros((N*n,3))
-    if n==1:
-        xi[:,0] = numpy.average(y[triangle[:],0], axis=1)
-        xi[:,1] = numpy.average(y[triangle[:],1], axis=1)
-        xi[:,2] = numpy.average(y[triangle[:],2], axis=1)
-
-    if n==3:
-        for i in range(N):
-            M = numpy.transpose(y[triangle[i]])
-            xi[n*i,:] = numpy.dot(M, numpy.array([0.5, 0.5, 0.]))
-            xi[n*i+1,:] = numpy.dot(M, numpy.array([0., 0.5, 0.5]))
-            xi[n*i+2,:] = numpy.dot(M, numpy.array([0.5, 0., 0.5]))
-
-    if n==4:
-        for i in range(N):
-            M = numpy.transpose(y[triangle[i]])
-            xi[n*i,:] = numpy.dot(M, numpy.array([1/3., 1/3., 1/3.]))
-            xi[n*i+1,:] = numpy.dot(M, numpy.array([3/5., 1/5., 1/5.]))
-            xi[n*i+2,:] = numpy.dot(M, numpy.array([1/5., 3/5., 1/5.]))
-            xi[n*i+3,:] = numpy.dot(M, numpy.array([1/5., 1/5., 3/5.]))
-
-    if n==7:
-        for i in range(N):
-            M = numpy.transpose(y[triangle[i]])
-            xi[n*i+0,:] = numpy.dot(M, numpy.array([1/3.,1/3.,1/3.]))
-            xi[n*i+1,:] = numpy.dot(M, numpy.array([.797426985353087,.101286507323456,.101286507323456]))
-            xi[n*i+2,:] = numpy.dot(M, numpy.array([.101286507323456,.797426985353087,.101286507323456]))
-            xi[n*i+3,:] = numpy.dot(M, numpy.array([.101286507323456,.101286507323456,.797426985353087]))
-            xi[n*i+4,:] = numpy.dot(M, numpy.array([.059715871789770,.470142064105115,.470142064105115]))
-            xi[n*i+5,:] = numpy.dot(M, numpy.array([.470142064105115,.059715871789770,.470142064105115]))
-            xi[n*i+6,:] = numpy.dot(M, numpy.array([.470142064105115,.470142064105115,.059715871789770]))
-
-    return xi[:,0], xi[:,1], xi[:,2]
