### 2016-08-01 c9f515c
format fixes and simplify preconditioner code
also one missed `self.`
