+++ pygbe/classes.py
@@ -4,9 +4,8 @@ import time
 from pygbe.tree.FMMutils import addSources, sortPoints, generateTree, findTwigs
 from pygbe.tree.direct import computeDiagonal
 from pygbe.util.semi_analytical import GQ_1D
-from pygbe.util.readData import (readVertex, readTriangle, readpqr, readcrd,
-                           readFields, read_surface)
 from pygbe.quadrature import quadratureRule_fine
+
 
 class Event():
     """
@@ -134,108 +133,21 @@ class Surface():
     """
 
     def __init__(self):
-#        self.triangle = []  # indices to triangle vertices
-#        self.vertex   = []  # position of vertices
-#        self.XinV     = []  # weights input for single layer potential
-#        self.XinK     = []  # weights input for double layer potential
-#        self.Xout_int = []  # output vector of interior operators
-#        self.Xout_ext = []  # output vector of exterior operators
-#        self.xi       = []  # x component of center
-#        self.yi       = []  # y component of center
-#        self.zi       = []  # z component of center
-#        self.xj       = []  # x component of gauss nodes
-#        self.yj       = []  # y component of gauss nodes
-#        self.zj       = []  # z component of gauss nodes
-#        self.Area     = []  # Area of triangles
-#        self.normal   = []  # normal of triangles
-#        self.sglInt_int = []  # singular integrals for V for internal equation
-#        self.sglInt_ext = []  # singular integrals for V for external equation
-#        self.xk       = []  # position of gauss points on edges
-#        self.wk       = []  # weight of gauss points on edges
-#        self.Xsk      = []  # position of gauss points for near singular integrals
-#        self.Wsk      = []  # weight of gauss points for near singular integrals
-#        self.tree     = []  # tree structure
-#        self.twig     = []  # tree twigs
-#        self.xiSort   = []  # sorted x component of center
-#        self.yiSort   = []  # sorted y component of center
-#        self.ziSort   = []  # sorted z component of center
-#        self.xjSort   = []  # sorted x component of gauss nodes
-#        self.yjSort   = []  # sorted y component of gauss nodes
-#        self.zjSort   = []  # sorted z component of gauss nodes
-#        self.xcSort   = []  # sorted x component box centers according to M2P_list array
-#        self.ycSort   = []  # sorted y component box centers according to M2P_list array
-#        self.zcSort   = []  # sorted z component box centers according to M2P_list array
-#        self.AreaSort = []  # sorted array of areas
-#        self.sglInt_intSort  = []  # sorted array of singular integrals for V for internal equation
-#        self.sglInt_extSort  = []  # sorted array of singular integrals for V for external equation
-#        self.unsort       = []  # array of indices to unsort targets
-#        self.triangleSort = []  # sorted array of triangles
-#        self.sortTarget   = []  # array of indices to sort targets
-#        self.sortSource   = []  # array of indices to sort sources
-#        self.offsetSource = []  # array with offsets to sorted source array
-#        self.offsetTarget = []  # array with offsets to sorted target array
-#        self.sizeTarget   = []  # array with number of targets per twig
-#        self.offsetTwigs  = []  # offset to twig in P2P list array
-#        self.P2P_list     = []  # pointers to twigs for P2P interaction list
[…]
