### 2016-08-03 #43 @kyleniemeyer
Alright, so I checked off the various items that are fine, but I have a few comments—and ran into some problems when trying to run the regression and performance tests.

1. The README is quite extensive and helpful (especially the instructions on building dependencies), but none of this information (installation, usage, examples, etc.) is actually contained in the documentation (http://barbagroup.github.io/pygbe/). I would recommend moving/repeating this information there, where it can be organized a bit more easily.

2. The paper is missing an archive/DOI of the version submitted (0.2), from (e.g.) Zenodo or figshare.

3. There are regression tests, but no unit tests (from what I can tell). This somewhat conflicts with the guidelines given in CONTRIBUTING.md, which says any new features need unit tests.

4. The support for CUDA versions below 7.0 seems somewhat restrictive, since we're on 7.5 now. I wasn't able to run the GPU-enabled version on my systems because of this. Would it be particularly challenging to support CUDA/nvcc 7.x?

I also have some comments on the paper:

1. There are some bugs with the affiliations for authors 1–3.
2. The summary could be a bit more general—as written, it isn't clear who would use the software and for what broad purpose.
3. The third reference is missing a DOI
4. You might consider adding references for the papers that have already used PyGBe (given in the README).

Regarding the tests, as I mentioned I couldn't run the GPU-enabled version of PyGBe, but it seems that I should have been able to run it in CPU-only mode. However, I ran into some errors with both the performance and regression tests, following the instructions. Both downloaded the external mesh files just fine, and seemed to go through some work (although I got frequent error messages and traceback for `pytools.prefork.ExecError: error invoking 'nvcc --version': [Errno 2] No such file or directory`).

The performance tests did not complete, with the error

    Traceback (most recent call last):
      File "run_and_plot_lysozome_case.py", line 249, in <module>
        main()
      File "run_and_plot_lysozome_case.py", line 245, in main
        generate_plot(compiled_results, filetype='pdf')
      File "run_and_plot_lysozome_case.py", line 85, in generate_plot
        pyg_ext = richardson_extrapolation(res)
      File "run_and_plot_lysozome_case.py", line 60, in richardson_extrapolation
        esolv = compiled_results['E_solv_kJ']
    KeyError: 'E_solv_kJ'

and the regression tests also led to an error:

    Traceback (most recent call last):
      File "run_all_regression_tests.py", line 26, in <module>
        except FileNotFoundError:
    NameError: name 'FileNotFoundError' is not defined

Finally, some minor comments:

Minor comments:

 * It might be nice to give a recipe to create a conda environment, for those of us that mainly use Python 3 😃
 * Running the regression (and I assume performance) tests requires matplotlib, which wasn't mentioned as a dependency
 * I noticed that the paper author list doesn't completely match the authors listed in the LICENSE, which may or may not be important.
