### 2016-08-03 #43 @gforsyth
Thanks, @kyleniemeyer !  We'll get on fixes for the pain points.
PyCUDA didn't support Cuda 7.5 last time we looked at upgrading, but it does now so we'll look into upgrading things.
