### 2016-08-03 #43 @ncclementi
Thanks, @kyleniemeyer for all the comments, we'll work on that.

About the errors that you're getting when you try to run the tests on CPU:
 Did you notice that you need to change all the `.param` files on the `input files` folder to set `GPU=0` (which force PyGBe to run on CPU)?
This is explained at the end of the [README_input_format](https://github.com/barbagroup/pygbe/blob/master/README_input_format).

However, I would strongly recommend to run the regression tests on the GPU because they take a while, running them on the CPU version will take even longer.
