### 2016-08-03 8795f05
Update paper.bib - Add DOI to YoonLenhoff1990
