### 2016-08-03 c1dcfe7
Update paper.md Fixing indentation
The orcid entry for the firs 3 authors was with the wrong indentation, this will probably fix the bug in the compiled paper.
