+++ paper.md
@@ -6,13 +6,13 @@ tags:
   - Poisson-Boltzmann
 authors:
  - name: Christopher D. Cooper
- - orcid: 0000-0003-0282-8998
+   orcid: 0000-0003-0282-8998
    affiliation: Universidad Técnica Federico Santa María
  - name: Natalia C. Clementi
- - orcid: 0000-0002-0575-5520
+   orcid: 0000-0002-0575-5520
    affiliation: The George Washington University
  - name: Gilbert Forsyth
- - orcid: 0000-0002-4983-1978
+   orcid: 0000-0002-4983-1978
    affiliation: The George Washington University
  - name: Lorena A. Barba
    orcid: 0000-0001-5812-2711
