### 2016-08-03 7e655e3
catch KeyError on performance runs if something fails
