### 2016-08-03 8fece21
script to simplify merging docs
