### 2016-08-03 1f99f26
update guides for better RST conversion
