### 2016-08-03 c7af783
update toctree in index.rst and fix helper script
