### 2016-08-03 2ea0d68
fix a few docstrings for RST formatting
