### 2016-08-03 a1ab65a
update input format readme
