### 2016-08-03 9f27f9b
IOError on Python2, not FileNotFoundError
I am going to be so happy to leave Python 2 behind.
