### 2016-08-03 c98b272
make pygbe run on CPU if no GPU available
no matter what it says in the param file
