### 2016-08-03 648de42
Added bib entries of papers using PyGBe
