+++ paper.bib
@@ -94,3 +94,41 @@
    issn = "0167-8191",
    doi = "10.1016/j.parco.2011.09.001",
 }
+
+@article{CooperClementiBarba2015,
+  title={Probing protein orientation near charged nanosurfaces for simulation-assisted biosensor design},
+  author={Cooper, Christopher D and Clementi, Natalia C and Barba, Lorena A},
+  journal={J. Chem. Phys.},
+  volume={143},
+  number={12},
+  pages={124709},
+  year={2015},
+  note = {Preprint on \href{arxiv:1503.08150}{arxiv:1503.08150}},
+  publisher={AIP Publishing}
+}
+
+@article{CooperBardhanBarba2014,
+    Author = {Cooper, Christopher D. and Bardhan, Jaydeep P. and Barba, L.~A.},
+    Date-Added = {2013-12-04 20:15:44 +0000},
+    Date-Modified = {2015-05-03 23:44:12 +0000},
+    Doi = {10.1016/j.cpc.2013.10.028},
+    Journal = {Comput. Phys. Commun.},
+    Keywords = {pygbe, Poisson-Boltzmann, BEM, barbagroup},
+    Month = {March},
+    Note = {Preprint on \href{http://arxiv.org/abs/1309.4018}{arXiv:/1309.4018}},
+    Number = {3},
+    Pages = {720--729},
+    Title = {A biomolecular electrostatics solver using {Python}, {GPU}s and boundary elements that can handle solvent-filled cavities and {S}tern layers.},
+    Volume = {185},
+    Year = {2014},
+    Bdsk-Url-1 = {http://dx.doi.org/10.1016/j.cpc.2013.10.028}}
+
+@article{CooperBarba2016,
+  title={Poisson--Boltzmann model for protein--surface electrostatic interactions and grid-convergence study using the PyGBe code},
+  author={Cooper, Christopher D and Barba, Lorena A},
+  journal={Computer Physics Communications},
+  volume={202},
+  pages={23--32},
+  year={2016},
+  publisher={Elsevier}
+}
