### 2016-08-03 1b87021
update dependencies and license copyright
