### 2016-08-03 724838a
Added info on applications using PyGBe
* References to papers using PyGBe
* Comments on applications where PyGBe is relevant
