### 2016-08-03 903efc6
Removed references to protein orientation papers
* Not included in this release
