### 2016-08-03 789d572
add quick conda command for easy install
