### 2016-08-03 f10da76
two small typos
