### 2016-08-03 5d3ca77
very small change for better sphinx docs
make `Run PyGBe` a top-level item in the sphinx build so that it gets
its own link on the sidebar
