### 2016-08-04 #43 @arfon
> @arfon, could we prevail upon you to regenerate the PDF?

Your wish... [10.21105.joss.00043.pdf](https://github.com/openjournals/joss-reviews/files/402384/10.21105.joss.00043.pdf)
