### 2016-08-04 04202c2
Fixing bibtex typo
