+++ paper.bib
@@ -13,7 +13,7 @@
 	Author = {B J. Yoon and A. M. Lenhoff},
 	Date-Added = {2016-07-13 19:06:52 +0000},
 	Date-Modified = {2016-07-13 19:06:52 +0000},
-	Doi = {10.1002/jcc.540110911}
+	Doi = {10.1002/jcc.540110911},
 	Journal = {J. Comput.\ Chem.},
 	Number = {9},
 	Pages = {1080--1086},
