### 2016-08-05 #43 @kyleniemeyer
Great improvements! The performance and regression tests now run fine for me, and the paper changes look good.

I did run into an issue with the performance test: I was running this on a headless server, and when the test finished I got an error that seems to be the code attempting to display something:
```
Traceback (most recent call last):
  File "run_and_plot_lysozome_case.py", line 254, in <module>
    main()
  File "run_and_plot_lysozome_case.py", line 250, in main
    generate_plot(compiled_results, filetype='pdf')
  File "run_and_plot_lysozome_case.py", line 93, in generate_plot
    pyplot.figure(figsize=(3, 2), dpi=80)
  File "/home/kylen/anaconda3/envs/python2/lib/python2.7/site-packages/matplotlib/pyplot.py", line 527, in figure
    **kwargs)
  File "/home/kylen/anaconda3/envs/python2/lib/python2.7/site-packages/matplotlib/backends/backend_qt4agg.py", line 46, in new_figure_manager
    return new_figure_manager_given_figure(num, thisFig)
  File "/home/kylen/anaconda3/envs/python2/lib/python2.7/site-packages/matplotlib/backends/backend_qt4agg.py", line 53, in new_figure_manager_given_figure
    canvas = FigureCanvasQTAgg(figure)
  File "/home/kylen/anaconda3/envs/python2/lib/python2.7/site-packages/matplotlib/backends/backend_qt4agg.py", line 76, in __init__
    FigureCanvasQT.__init__(self, figure)
  File "/home/kylen/anaconda3/envs/python2/lib/python2.7/site-packages/matplotlib/backends/backend_qt4.py", line 68, in __init__
    _create_qApp()
  File "/home/kylen/anaconda3/envs/python2/lib/python2.7/site-packages/matplotlib/backends/backend_qt5.py", line 138, in _create_qApp
    raise RuntimeError('Invalid DISPLAY variable')
RuntimeError: Invalid DISPLAY variable
```
I assume this wouldn't be a problem if this was on a local machine. However, the behavior I would expect is to create the plot (replacing `time_lys.png` I assume) regardless of whether something can be shown on screen.
