### 2016-08-05 #43 @kyleniemeyer
Lastly, it looks like the documentation site (http://barbagroup.github.io/pygbe/) is now down.
