### 2016-08-05 0bdb925
make matplotlib use non-X-dependent backend
