### 2016-08-05 #43 @gforsyth
> I did run into an issue with the performance test: I was running this on a headless server, and when the test finished I got an error that seems to be the code attempting to display something

Ahh, yeah, X-dependent backend by default.  This should be fixed on master now but I'm running it on a headless server to confirm.

> Lastly, it looks like the documentation site (http://barbagroup.github.io/pygbe/) is now down.

Yikes.  Not sure what happened there, but it's back up now.  Thanks!
