### 2016-08-05 #43 @arfon
> Awesome, that worked perfectly for me.

👍

@gforsyth - could you add a link to the software archive (e.g. Zenodo/figshare) in this thread? I can then move forward with accepting this submission.
