### 2016-08-05 #43 @gforsyth
Hey @arfon, I'm happy to.  Quick question re protocol:  Is it better to do a minor version bump with the post-review changes and use that or should I delete and re-tag and then submit that to Zenodo?
