### 2016-08-05 #43 @arfon
> Hey @arfon, I'm happy to. Quick question re protocol: Is it better to do a minor version bump with the post-review changes and use that or should I delete and re-tag and then submit that to Zenodo?

I think the minor version bump is probably better.
