### 2016-08-05 26802a4
update changelog for minor release bump
