### 2016-08-05 #43 @keith923
I'm afraid that I can't get PyCuda to build on OS X El Capitan.  It's looking for 64 bit CUDA libs, and they don't exist in CUDA 7.5, or 6.0.  I'd need to switch to a Linux VM, and I'm not sure that CUDA will properly use the GPU in there.
