### 2016-08-05 #43 @gforsyth
@keith923 -- possible fix for El Capitan and PyCUDA (sorry, no mac to test it on here):

```console
python configure.py --cuda-root=/usr/local/cuda --ldflags="-F/Library/Frameworks -framework CUDA" --cxxflags="-arch x86_64"
make
sudo make install
```
