### 2016-08-08 #43 @keith923
Thanks @gforsyth.  That did help, and I ran into another error that made me look at the problem a bit harder.  Short story is that NVIDIA have a prefpane for updating the CUDA driver, but one needs to update the SDK manually.  I hopefully won't be much longer.
