+++ tests/test_lysozyme.py
@@ -1,0 +1,5 @@
+from pygbe.main import main
+
+def test_lysozyme():
+    res = main(['', '../examples/lys', return_output_dict=True])
+
