### 2016-08-09 f3de359
remove debugging sleep command
yikes, not a good thing to have in there...
