### 2016-08-09 4319e91
fix generator steps, regex decoding and rm sleep
damn sleep command is stupid
