### 2016-08-09 #43 @lizutah
Conflict of interest

General checks
- [x]  Repository: Is the source code for this software available at the repository url?
- [x]  License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
- [x]   Version: Does the release version given match the GitHub release (0.2)?

**Mostly. All the examples worked. @keith923 is posting errors he found in the regression tests**
- [x] Performance: Have any performance claims of the software been confirmed?
**Yes, The GPU performance is obvious and the code executes quickly.**

**Yes.  Pointers to tests and examples exist in the source tree.**
- [x] Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g. API method documentation)?
- [x]  Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

**Some notes:**
1. Because the code is in github, it’s pretty clear how users can contribute to the software.  However there are instructions for contribution.
2. I do not see a mailing list or contact for questions, although github provides a mechanism for this
3. Same as 2.

Other:
“API documentation” isn’t being used correctly.  This is really the “User Guide.”  The term “API Guide” is used in installation.htmp and in the README.md

The actual API should be a list of modules and description of the code (more like the “Indices and Tables”).  You could also call this the “Developer Documentation.”

“Show source” link is broken on left side of page.

Additionally http://barbagroup.github.io/pygbe/ needs a home button.  Once you start clicking on links, it’s hard to get back to the main page and the menu in the left changes so you forget how to get back to where you came from.**

- [x] Authors: Does the paper.md file include a list of authors with their affiliations?
**Seems like the affiliations could be combined when they are the same?**
- [x] References: Do all archival references that should have a DOI list one (e.g. papers, datasets, software)?
