### 2016-08-09 #43 @keith923
Running the regression tests I found the following errors.

The following tests:

* Config file: ./pygbe/tests/regression_tests/input_files/lys_8.config
Parameter file: ./pygbe/tests/regression_tests/input_files/lys.param

* Config file: ./pygbe/tests/regression_tests/input_files/molecule_stern_32K.config
Parameter file: ./pygbe/tests/regression_tests/input_files/sphere_fine.param

* Config file: ./pygbe/tests/regression_tests/input_files/neumann_surface_32K.config
Parameter file: ./pygbe/tests/regression_tests/input_files/sphere_fine.param

* Config file: ./pygbe/tests/regression_tests/input_files/twosphere_dirichlet_32K.config
Parameter file: ./pygbe/tests/regression_tests/input_files/sphere_fine.param

* Config file: ./pygbe/tests/regression_tests/input_files/twosphere_neumann_32K.config
Parameter file: ./pygbe/tests/regression_tests/input_files/sphere_fine.param

* Config file: ./pygbe/tests/regression_tests/input_files/twosphere_32K.config
Parameter file: ./pygbe/tests/regression_tests/input_files/sphere_fine.param

erred with:
```
Traceback (most recent call last):
  File "lysozyme.py", line 123, in <module>
    main()
  File "lysozyme.py", line 33, in main
    N, iterations, Esolv, Esurf, Ecoul, Time = run_regression(mesh, test_name, problem_folder, param)
  File "/Users/d3x923/Documents/projects/pygbe/tests/regression_tests/regression.py", line 133, in run_regression
    '{}'.format(problem_folder),], return_output_fname=True)
  File "/usr/local/lib/python2.7/site-packages/PyGBe-0.2-py2.7-macosx-10.11-x86_64.egg/pygbe/main.py", line 340, in main
    dataTransfer(surf_array, field_array, ind0, param, kernel)
  File "/usr/local/lib/python2.7/site-packages/PyGBe-0.2-py2.7-macosx-10.11-x86_64.egg/pygbe/gpuio.py", line 58, in dataTransfer
    s].ycSort.astype(REAL)))
  File "/usr/local/lib/python2.7/site-packages/pycuda/gpuarray.py", line 975, in to_gpu
    result = GPUArray(ary.shape, ary.dtype, allocator, strides=_compact_strides(ary))
  File "/usr/local/lib/python2.7/site-packages/pycuda/gpuarray.py", line 209, in __init__
    self.gpudata = self.allocator(self.size * self.dtype.itemsize)
pycuda._driver.MemoryError: cuMemAlloc failed: out of memory
```

And these tests:

* Config file: ./pygbe/tests/regression_tests/input_files/molecule_dirichlet_32K.config
Parameter file: ./pygbe/tests/regression_tests/input_files/sphere_fine.param

* Config file: ./pygbe/tests/regression_tests/input_files/molecule_neumann_32K.config
Parameter file: ./pygbe/tests/regression_tests/input_files/sphere_fine.param

* Config file: ./pygbe/tests/regression_tests/input_files/dirichlet_surface_32K.config
Parameter file: ./pygbe/tests/regression_tests/input_files/sphere_fine.param

* Config file: ./pygbe/tests/regression_tests/input_files/molecule_single_32K.config
Parameter file: ./pygbe/tests/regression_tests/input_files/sphere_fine.param

erred with:

```
Traceback (most recent call last):
  File "molecule_dirichlet.py", line 76, in <module>
    main()
  File "molecule_dirichlet.py", line 28, in main
    mesh, test_name, problem_folder, param)
  File "/Users/d3x923/Documents/projects/pygbe/tests/regression_tests/regression.py", line 133, in run_regression
    '{}'.format(problem_folder),], return_output_fname=True)
  File "/usr/local/lib/python2.7/site-packages/PyGBe-0.2-py2.7-macosx-10.11-x86_64.egg/pygbe/main.py", line 353, in main
    ind0)
  File "/usr/local/lib/python2.7/site-packages/PyGBe-0.2-py2.7-macosx-10.11-x86_64.egg/pygbe/matrixfree.py", line 900, in generateRHS_gpu
    ind0, timing, kernel)
  File "/usr/local/lib/python2.7/site-packages/PyGBe-0.2-py2.7-macosx-10.11-x86_64.egg/pygbe/projection.py", line 142, in project
    param, LorY, timing, kernel)
  File "/usr/local/lib/python2.7/site-packages/PyGBe-0.2-py2.7-macosx-10.11-x86_64.egg/pygbe/tree/FMMutils.py", line 901, in M2P_gpu
    toc.synchronize()
pycuda._driver.LaunchError: cuEventSynchronize failed: the launch timed out and was terminated
```
