### 2016-08-09 #43 @labarba
@keith923 : What CUDA card are you using?
