### 2016-08-09 #43 @gforsyth
Thanks, @lizutah -- I'm working on fixing up the docs as you suggested.

@keith923 -- those all stem from memory issues, I believe.  The cards we test on all have 5+ GB of memory and I think the larger meshes are causing problems on your card.

If you edit `tests/regression_tests/regression.py` as follows it should prevent those errors.  Sorry for the trouble!

```console
diff --git a/tests/regression_tests/regression.py b/tests/regression_tests/regression.py
index db05284..08c7553 100644
--- a/tests/regression_tests/regression.py
+++ b/tests/regression_tests/regression.py
@@ -14,8 +14,8 @@ ESURF_REGEX = re.compile('[^: ]Esurf = (\-*\d*\.\d*)\ kcal\/mol')
 ECOUL_REGEX = re.compile('[^: ]Ecoul = (\-*\d*\.\d*)\ kcal\/mol')
 TIME_REGEX = re.compile('Time = (\-*\d*\.\d*)\ s')

-mesh = ['500', '2K', '8K', '32K']
-lysozome_mesh = ['1','2','4','8']
+mesh = ['500', '2K', '8K']
+lysozome_mesh = ['1','2','4']

 def picklesave(test_outputs):
     with open('tests','w') as f:
```
