### 2016-08-09 #43 @keith923
@labarba my Mac has a GeForce GT 750M with only 2GB of RAM.
@gforsyth while it makes sense to not try to run with the finer meshes, it might be nice if the software caught the exception and reported the problem to the user.  If it were me, I might even consider printing a warning and then using a coarser mesh.  Perhaps after asking the user if they'd like to proceed like that.
