### 2016-08-09 #43 @gforsyth
@keith923 -- most definitely.  I have a patch in the works to do just that.
