### 2016-08-11 b8bd90f
ignore problem files when building docs
also remove show page source links, since they don't work with napoleon
