### 2016-08-11 30ad8da
rename API docs to developer info
