### 2016-08-11 293ff42
update title heading for index page in docs
