### 2016-08-11 1a5a6df
add .travis.yml with doctr config to push docs
