+++ .travis.yml
@@ -1,0 +1,31 @@
+language: python
+python:
+  - "3.5"
+env:
+  global:
+    - secure: "bLP+5HW9seXOSi1Kyjf1XyOiLBOWf/hDKIW5f+/TCi5afFGdxlQp4BNcxwG9ZQ61Er2dcD5kuQ3Lop+AA3GSpt43a3cPhCZ/LX37b17QKeEAr8IhzgrAoH+hlQheWnbgzGxvOYjhywUwIU+bsotRoZLRrrwBYZXqOWLshFCeReM="
+
+
+sudo: false
+
+# install miniconda and create env for pygbe
+# also use pip to install doctr py3.5
+install:
+  - pip install doctr
+  - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh;
+  - bash miniconda.sh -b -p $HOME/miniconda
+  - export PATH="$HOME/miniconda/bin:$PATH"
+  - hash -r
+  - conda config --set always_yes yes --set changeps1 no
+  - conda create -q -n pygbe python=2.7 numpy scipy swig sphinx
+  - source activate pygbe
+
+script:
+  - sphinx-apidoc -f -o docs/source pygbe
+  - cd docs
+  - python update_index.py
+  - make html
+  - cd ..
+  - source deactivate
+  - python -m doctr deploy
+
