### 2016-08-11 7189260
install pandoc in travis build
