### 2016-08-11 7843a77
rename miniconda download package
