+++ .travis.yml
@@ -15,7 +15,7 @@ addons:
 # also use pip to install doctr py3.5
 before_install:
   - pip install doctr
-  - wget https://repo.continuum.io/miniconda/Miniconda2-latest-Linux-x86_64.sh
+  - wget https://repo.continuum.io/miniconda/Miniconda2-latest-Linux-x86_64.sh -O miniconda.sh;
   - bash miniconda.sh -b -p $HOME/miniconda
   - export PATH="$HOME/miniconda/bin:$PATH"
   - hash -r
