### 2016-08-11 6c07a35
install doctr inside base py3 install so it works
py2 env will be deactivated first
