+++ .travis.yml
@@ -14,11 +14,11 @@ addons:
 # install miniconda and create env for pygbe
 # also use pip to install doctr py3.5
 before_install:
-  - pip install doctr
-  - wget https://repo.continuum.io/miniconda/Miniconda2-latest-Linux-x86_64.sh -O miniconda.sh;
+  - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh;
   - bash miniconda.sh -b -p $HOME/miniconda
   - export PATH="$HOME/miniconda/bin:$PATH"
   - hash -r
+  - pip install doctr
   - conda config --set always_yes yes --set changeps1 no
   - conda create -q -n pygbe python=2.7 numpy scipy swig sphinx
   - source activate pygbe
