+++ docs/_sources/source/pygbe.util.txt
@@ -1,0 +1,38 @@
+pygbe.util package
+==================
+
+Submodules
+----------
+
+pygbe.util.an_solution module
+-----------------------------
+
+.. automodule:: pygbe.util.an_solution
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+pygbe.util.readData module
+--------------------------
+
+.. automodule:: pygbe.util.readData
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+pygbe.util.semi_analytical module
+---------------------------------
+
+.. automodule:: pygbe.util.semi_analytical
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: pygbe.util
+    :members:
+    :undoc-members:
+    :show-inheritance:
