+++ docs/_sources/source/pygbe.tree.txt
@@ -1,0 +1,30 @@
+pygbe.tree package
+==================
+
+Submodules
+----------
+
+pygbe.tree.FMMutils module
+--------------------------
+
+.. automodule:: pygbe.tree.FMMutils
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+pygbe.tree.cuda_kernels module
+------------------------------
+
+.. automodule:: pygbe.tree.cuda_kernels
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: pygbe.tree
+    :members:
+    :undoc-members:
+    :show-inheritance:
