+++ docs/_sources/index.txt
@@ -1,0 +1,25 @@
+.. PyGBe documentation master file, created by
+   sphinx-quickstart on Wed Jun 22 15:34:00 2016.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+.. include:: intro.rst
+
+.. toctree::
+   :titlesonly:
+   :maxdepth: 2
+
+   installation
+   contributing
+   input_format
+   references
+
+
+
+Developer Documentation
+=======================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
+

[…]
