### 2016-08-12 8fb5f76
update link to docs in README
For now, the gh-pages address needs to have `docs` appended to it so we
can use doctr to automatically generate the documentation
