### 2016-08-12 c5b3c9b
update documentation instructions in contributing
no longer required to commit any docs on user side, this is all handled
automatically by travis and doctr
