### 2016-08-12 f169643
Update docs after building Travis build 5
