### 2016-08-12 c76c376
add explicit install step to travis.yml
I believe this is required to make sure that all the C-dependency stuff
actually builds and so is then available for documentation generation.
