+++ .travis.yml
@@ -24,6 +24,7 @@ before_install:
   - source activate pygbe
 
 script:
+  - python setup.py install
   - sphinx-apidoc -f -o docs/source pygbe
   - cd docs
   - python update_index.py
