### 2016-08-12 b7c1216
Update docs after building Travis build 6
