+++ docs/_sources/source/pygbe.tree.txt
@@ -12,10 +12,34 @@ pygbe.tree.FMMutils module
     :undoc-members:
     :show-inheritance:
 
+pygbe.tree.calculateMultipoles module
+-------------------------------------
+
+.. automodule:: pygbe.tree.calculateMultipoles
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
 pygbe.tree.cuda_kernels module
 ------------------------------
 
 .. automodule:: pygbe.tree.cuda_kernels
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+pygbe.tree.direct module
+------------------------
+
+.. automodule:: pygbe.tree.direct
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+pygbe.tree.multipole module
+---------------------------
+
+.. automodule:: pygbe.tree.multipole
     :members:
     :undoc-members:
     :show-inheritance:
