### 2016-08-12 23aeac9
add 'home'ish button to documentation
As requested in review #2
