+++ docs/index.rst
@@ -1,9 +1,9 @@
-.. PyGBe documentation master file, created by
-   sphinx-quickstart on Wed Jun 22 15:34:00 2016.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
 .. include:: intro.rst
+
+.. toctree::
+   :hidden:
+
+   self
 
 .. toctree::
    :titlesonly:
