### 2016-08-12 c6d27fe
Update docs after building Travis build 9
