+++ docs/_sources/index.txt
@@ -1,7 +1,6 @@
 .. include:: intro.rst
 
 .. toctree::
-   :hidden:
 
    self
 
