### 2016-08-12 7d8b29b
add explicit sidebar links to docs
