+++ docs/conf.py
@@ -157,7 +157,13 @@ html_static_path = ['_static']
 #html_use_smartypants = True
 
 # Custom sidebar templates, maps document names to template names.
-#html_sidebars = {}
+html_sidebars = {
+    '**': [
+        'about.html',
+        'navigation.html',
+        'relations.html',
+        'searchbox.html',
+    ]}
 
 # Additional templates that should be rendered to pages, maps page names to
 # template names.
