### 2016-08-12 bfe2002
Update side-toolbar with better sphinx options
Removed the explicit link at the top in favor of the template-driven
version that also includes the project name.  Should also (hopefully)
include a button linking back to the github repo.
