+++ docs/conf.py
@@ -117,7 +117,12 @@ html_theme = 'alabaster'
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
 # documentation.
-#html_theme_options = {}
+html_theme_options = {
+    'description': 'Python, GPUs and Boundary elements for biomolecular electrostatics',
+    'github_user': 'barbagroup',
+    'github_repo': 'pygbe',
+    'github_button': 'true',
+}
 
 # Add any paths that contain custom themes here, relative to this directory.
 #html_theme_path = []
