### 2016-08-12 9f67835
Update docs after building Travis build 11
