+++ docs/_sources/index.txt
@@ -1,10 +1,6 @@
 .. include:: intro.rst
 
 .. toctree::
-
-   self
-
-.. toctree::
    :titlesonly:
    :maxdepth: 2
 
