### 2016-08-12 #43 @gforsyth
@lizutah -- I've pushed documentation fixes as you suggested.

@keith923 -- we now catch both out of memory and synchronize errors with human readable error messages

Thanks both for your comments and please let us know if there's anything else that needs doing.
