### 2016-08-16 23f0b9f
add kwarg to directly return results dictionary
