+++ pygbe/main.py
@@ -192,7 +192,8 @@ def check_nvcc_version():
                  'or remove `nvcc` from your PATH to use CPU only.')
 
 
-def main(argv=sys.argv, log_output=True, return_output_fname=False):
+def main(argv=sys.argv, log_output=True, return_output_fname=False,
+         return_results_dict=False):
     """
     Run a PyGBe problem, write outputs to STDOUT and to log file in
     problem directory
@@ -446,6 +447,9 @@ def main(argv=sys.argv, log_output=True, return_output_fname=False):
     #reset stdout so regression tests, etc, don't get logged into the output
     #file that they themselves are trying to read
     sys.stdout = sys.__stdout__
+
+    if return_results_dict:
+        return results_dict
 
     if return_output_fname and log_output:
         return outputfname
