### 2016-08-16 f1eacc7
add pickle results and pytest regression
