### 2016-08-16 c821030
specify protocol for using pickle
this ensures that results generated using python2 can be read by python3
and vice-versa.  this will probably not be necessary after we finish
porting everything
