+++ pygbe/main.py
@@ -442,7 +442,7 @@ def main(argv=sys.argv, log_output=True, return_output_fname=False,
     output_pickle.append('resultspickle')
     output_pickle = '-'.join(output_pickle)
     with open(os.path.join(output_dir, output_pickle), 'wb') as f:
-        pickle.dump(results_dict, f)
+        pickle.dump(results_dict, f, 2)
 
     #reset stdout so regression tests, etc, don't get logged into the output
     #file that they themselves are trying to read
