### 2016-08-16 7050be9
fix lysozyme test
