### 2016-08-16 8aa2184
add sphere regression test
