### 2016-08-16 2fd47e0
fix print statement for pgbmut regression test
