### 2016-08-17 9f21e34
rename all old regression tests as convergence tests
because they are convergence tests
