+++ tests/convergence_tests/twosphere_neumann.py
@@ -5,7 +5,7 @@ import numpy
 import pickle
 
 from pygbe.util import an_solution
-from regression import (scanOutput, run_regression, picklesave, pickleload,
+from convergence import (scanOutput, run_convergence, picklesave, pickleload,
                         report_results, mesh)
 
 
@@ -24,7 +24,7 @@ def main():
     param = 'sphere_fine.param'
     test_name = 'twosphere_neumann'
     if test_name not in test_outputs.keys():
-        N, iterations, Esolv, Esurf, Ecoul, Time = run_regression(
+        N, iterations, Esolv, Esurf, Ecoul, Time = run_convergence(
             mesh, test_name, problem_folder, param)
         test_outputs[test_name] = [N, iterations, Esolv, Esurf, Ecoul, Time]
 
@@ -35,7 +35,7 @@ def main():
     param = 'sphere_fine.param'
     test_name = 'neumann_surface'
     if test_name not in test_outputs.keys():
-        N, iterations, Esolv, Esurf, Ecoul, Time = run_regression(
+        N, iterations, Esolv, Esurf, Ecoul, Time = run_convergence(
             mesh, test_name, problem_folder, param)
         test_outputs[test_name] = [N, iterations, Esolv, Esurf, Ecoul, Time]
 
