+++ tests/convergence_tests/sphere_molecule_stern.py
@@ -5,7 +5,7 @@ import numpy
 import pickle
 
 from pygbe.util import an_solution
-from regression import (scanOutput, run_regression, picklesave, pickleload,
+from convergence import (scanOutput, run_convergence, picklesave, pickleload,
                         report_results, mesh)
 
 
@@ -22,7 +22,7 @@ def main():
     param = 'sphere_fine.param'
     test_name = 'molecule_stern'
     if test_name not in test_outputs.keys():
-        N, iterations, Esolv, Esurf, Ecoul, Time = run_regression(
+        N, iterations, Esolv, Esurf, Ecoul, Time = run_convergence(
             mesh, test_name, problem_folder, param)
         test_outputs[test_name] = [N, iterations, Esolv, Esurf, Ecoul, Time]
 
