### 2016-08-17 453cb86
update contributing docs with new docs/tests info
Documentation is handled by the doctr, so no input is required from
users unless they need to check changes locally (probably a good idea).

Test info updated to reflect the new, faster regression tests and to
explain that the convergence tests are also available.
