### 2016-08-17 0dd7ae2
Fixing print statement
