### 2016-08-17 23ae3fb
Fixing import, missing readElectricField function
