### 2016-08-17 973917f
collapse average into single line
