+++ pygbe/classes.py
@@ -22,302 +22,6 @@ class Event():
         pass
 
 
-class Surface():
-    """
-    Surface class.
-    It contains information about the solvent excluded surface.
-
-    Attributes
-    -----------
-
-    triangle      : list, indices to triangle vertices.
-    vertex        : list, position of vertices.
-    XinV          : list, weights input for single layer potential.
-    XinK          : list, weights input for double layer potential.
-    Xout_int      : list, output vector of interior operators.
-    Xout_ext      : list, output vector of exterior operators.
-    xi            : list, x component of center.
-    yi            : list, y component of center.
-    zi            : list, z component of center.
-    xj            : list, x component of gauss nodes.
-    yj            : list, y component of gauss nodes.
-    zj            : list, z component of gauss nodes.
-    Area          : list, areas of triangles.
-    normal        : list, normal of triangles.
-    sglInt_int    : list, singular integrals for V for internal equation.
-    sglInt_ext    : list, singular integrals for V for external equation.
-    xk            : list, position of gauss points on edges.
-    wk            : list, weight of gauss points on edges.
-    Xsk           : list, position of gauss points for near singular integrals.
-    Wsk           : list, weight of gauss points for near singular integrals.
-    tree          : list, tree structure.
-    twig          : list, tree twigs.
-    xiSort        : list, sorted x component of center.
-    yiSort        : list, sorted y component of center.
-    ziSort        : list, sorted z component of center.
-    xjSort        : list, sorted x component of gauss nodes.
-    yjSort        : list, sorted y component of gauss nodes.
-    zjSort        : list, sorted z component of gauss nodes.
-    xcSort        : list, sorted x component of the box centers according to
-                          M2P_list array.
-    ycSort        : list, sorted y component of the box centers according to
-                          M2P_list array.
-    zcSort        : list, sorted z component of the box centers according to
-                          M2P_list array.
-    AreaSort      : list, sorted array of areas.
-    sglInt_intSort: list, sorted array of singular integrals for V for internal
-                          equation.
-    sglInt_extSort: list, sorted array of singular integrals for V for external
-                          equation.
-    unsort        : list, array of indices to unsort targets.
-    triangleSort  : list, sorted array of triangles.
-    sortTarget    : list, indices to sort targets.
-    sortSource    : list, indices to sort sources.
-    offsetSource  : list, offsets to sorted source array.
-    offsetTarget  : list, offsets to sorted target array.
-    sizeTarget    : list, number of targets per twig.
-    offsetTwigs   : list, offset to twig in P2P list array.
-    P2P_list      : list, pointers to twigs for P2P interaction list.
-    offsetMlt     : list, offset to multipoles in M2P list array.
-    M2P_list      : list, pointers to boxes for M2P interaction list.
-    Precond       : list, sparse representation of preconditioner for self
-                          interaction block.
-    Ein           : float, permitivitty inside surface.
-    Eout          : float, permitivitty outside surface.
-    E_hat         : float, ratio of Ein/Eout.
-    kappa_in      : float, kappa inside surface.
-    kappa_out     : float, kappa inside surface.
-    LorY_in       : int, Laplace (1) or Yukawa (2) in inner region.
[…]
