+++ pygbe/surface.py
@@ -9,10 +9,10 @@ from scipy import linalg
 from pygbe.tree.FMMutils import addSources, sortPoints, generateTree, findTwigs
 from pygbe.tree.direct import computeDiagonal
 from pygbe.util.semi_analytical import GQ_1D
-from pygbe.util.readData import (readVertex, readTriangle, readpqr, readcrd,
-                           readFields, read_surface)
 from pygbe.quadrature import quadratureRule_fine
-from pygbe.classes import Surface, Field
+from pygbe.util.readData import (readVertex, readTriangle, readpqr, readcrd,
+                                 readFields, read_surface)
+from pygbe.classes import Field
 
 
 def initialize_surface(field_array, filename, param):
@@ -54,7 +54,7 @@ def initialize_surface(field_array, filename, param):
         triangle_raw = readTriangle(files[i] + '.face', s.surf_type)
         toc = time.time()
         print('Time load mesh: {}'.format(toc - tic))
-        Area_null = zero_areas(s, triangle_raw, Area_null)
+        Area_null = s.zero_areas(triangle_raw, Area_null)
         s.triangle = numpy.delete(triangle_raw, Area_null, 0)
         print('Removed areas=0: {}'.format(len(Area_null)))
 
@@ -78,33 +78,6 @@ def initialize_surface(field_array, filename, param):
 
         surf_array.append(s)
     return surf_array
-
-
-def zero_areas(s, triangle_raw, Area_null):
-    """
-    Looks for "zero-areas", areas that are really small, almost zero. It appends
-    them to Area_null list.
-
-    Arguments
-    ----------
-    s           : class, surface where we whan to look for zero areas.
-    triangle_raw: list, triangles of the surface.
-    Area_null   : list, contains the zero areas.
-
-    Returns
-    --------
-    Area_null   : list, indices of the triangles with zero-areas.
-    """
-
-    for i in range(len(triangle_raw)):
-        L0 = s.vertex[triangle_raw[i, 1]] - s.vertex[triangle_raw[i, 0]]
-        L2 = s.vertex[triangle_raw[i, 0]] - s.vertex[triangle_raw[i, 2]]
-        normal_aux = numpy.cross(L0, L2)
-        Area_aux = linalg.norm(normal_aux) / 2
-        if Area_aux < 1e-10:
-            Area_null.append(i)
-
-    return Area_null
 
 
 def initialize_field(filename, param):
@@ -175,6 +148,7 @@ def initialize_field(filename, param):
 
 
 def fill_phi(phi, surf_array):
+
     """
     It places the result vector on surface structure.
 
@@ -205,3 +179,328 @@ def fill_phi(phi, surf_array):
             surf_array[i].phi = phi[s_start:s_start + s_size]
             surf_array[i].dphi = phi[s_start + s_size:s_start + 2 * s_size]
             s_start += 2 * s_size
+
+
+
+class Surface():
+    """
+    Surface class.
+    It contains information about the solvent excluded surface.
+
+    Attributes
+    -----------
+
+    triangle      : list, indices to triangle vertices.
+    vertex        : list, position of vertices.
+    XinV          : list, weights input for single layer potential.
+    XinK          : list, weights input for double layer potential.
+    Xout_int      : list, output vector of interior operators.
+    Xout_ext      : list, output vector of exterior operators.
+    xi            : list, x component of center.
+    yi            : list, y component of center.
+    zi            : list, z component of center.
+    xj            : list, x component of gauss nodes.
+    yj            : list, y component of gauss nodes.
+    zj            : list, z component of gauss nodes.
+    Area          : list, areas of triangles.
+    normal        : list, normal of triangles.
+    sglInt_int    : list, singular integrals for V for internal equation.
[…]
