### 2016-08-17 d7833cc
move surface def mechanics inside surface class
