### 2016-08-17 9b67dbc
drop case on attributes in surface class
