### 2016-08-18 #43 @arfon
@labarba - seems like we're ready to move forward on this 🎉

At this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
