### 2016-08-18 #43 @gforsyth
Version bump to 2.1 with all revisions.

10.5281/zenodo.60474

:tada::tada::tada:
