### 2016-08-18 18c4ded
make regression tests py3 compatible
