+++ tests/regression_tests/sphere_dirichlet.py
@@ -13,7 +13,7 @@ def main():
     print('{:-^60}'.format('Running sphere_dirichlet test'))
     try:
         test_outputs = pickleload()
-    except IOError:
+    except FileNotFoundError:
         test_outputs = {}
 
     problem_folder = 'input_files'
