+++ tests/regression_tests/lysozyme.py
@@ -12,7 +12,7 @@ def main():
     print('{:-^60}'.format('Running lysozyme test'))
     try:
         test_outputs = pickleload()
-    except IOError:
+    except FileNotFoundError:
         test_outputs = {}
 
     problem_folder = 'input_files'
