+++ tests/regression_tests/molecule_neumann.py
@@ -14,7 +14,7 @@ def main():
 
     try:
         test_outputs = pickleload()
-    except IOError:
+    except FileNotFoundError:
         test_outputs = {}
 
     problem_folder = 'input_files'
