+++ tests/regression_tests/twosphere_neumann.py
@@ -14,7 +14,7 @@ def main():
     print('{:-^60}'.format('Running twosphere_neumann test'))
     try:
         test_outputs = pickleload()
-    except IOError:
+    except FileNotFoundError:
         test_outputs = {}
 
     problem_folder = 'input_files'
