### 2016-08-18 #43 @arfon
Thanks for the review @kyleniemeyer!

@gforsyth @labarba your paper is now accepted and your DOI is http://dx.doi.org/10.21105/joss.00043 🚀 🎉 💥
